% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_composition.R
\name{compose_ofb}
\alias{compose_ofb}
\title{Mise en page de graphiques ggplot}
\usage{
compose_ofb(gg, src = NULL, src_size = 2, legend_dir = "vertical")
}
\arguments{
\item{gg}{un graphique obtenu avec le package \{ggplot2\}}

\item{src}{vecteur de texte, chaque élément correspondant à une source des
données utilisées pour produire le graphique}

\item{src_size}{taille de la police à utiliser pour les sources de données}

\item{legend_dir}{sens dans lequel les (si plusieurs) légendes sont arrangées
("horizontal" ou "vertical")}
}
\description{
Mise en page de graphiques obtenus avec \{ggplot2\}. La légende, la date
d'édition du graphique, les sources de données et le logo de l'OFB sont
placés à droite du graphique.
}
\examples{
\dontrun{
gg <- data.frame(
        x = rnorm(100),
        y = rnorm(100),
        c = sample(c("a", "b", "c"), size = 100, replace = TRUE),
        d = rnorm(100)
      ) \%>\%
    ggplot2::ggplot() +
    ggplot2::geom_point(ggplot2::aes(x = x, y = y, colour = c, size = d)) +
    ggplot2::labs(title = "titre", subtitle = "sous-titre")

gg
compose_ofb(
    gg,
    src = c("OFB", "partenaires"),
    legend_dir = "horizontal"
  )
}
}
