% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_note_ofb.R
\name{creer_note_ofb}
\alias{creer_note_ofb}
\title{Créer une note OFB}
\usage{
creer_note_ofb(nom, dossier = getwd())
}
\arguments{
\item{nom}{nom du fichier Quarto/Word à créer}

\item{dossier}{dossier où les fichiers seront crées}
}
\description{
Cette fonction permet de créer un fichier \href{https://quarto.org/}{Quarto} permettant de générer une note au format word selon le modèle de l'OFB.
}
