% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_template.R
\name{golem_template}
\alias{golem_template}
\title{Modèle OFB de tableau de bord Shiny}
\usage{
golem_template(path, package_name, ...)
}
\arguments{
\item{path}{chemin du dossier où créer le package de l'appli}

\item{package_name}{nom du package}

\item{...}{autres arguments passés à \code{\link[golem:create_golem]{golem::create_golem()}}}
}
\description{
Cette fonction permet de personnaliser les applis shiny créees à l'aide du
package \{golem\} pour générer un modèle de tableau de bord selon la charte
graphique de l'OFB
}
\examples{
\dontrun{
golem::create_golem(
      "TableauBord",
       project_hook = templates:golem_template
     )
}

}
\seealso{
\code{\link[golem:create_golem]{golem::create_golem()}}
}
