% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ofb_colors}
\alias{ofb_colors}
\title{Couleurs de la charte graphique de l'OFB}
\format{
vecteur des codes HEX des 18 couleurs de la charte graphique de l'OFB. Les noms de couleurs utilisés correspondent à la couleur principale (bleu, vert, marron, rouge, orange et jaune) suivie d'un numéro lorsque plusieurs nuances sont présentes. Les numéros sont croissants de la nuance la plus foncée à la plus claire
}
\usage{
ofb_colors
}
\description{
Couleurs de la charte graphique de l'OFB
}
\keyword{datasets}
