% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{ofb_cols}
\alias{ofb_cols}
\title{Récupérer les codes HEX des couleurs OFB}
\usage{
ofb_cols(...)
}
\arguments{
\item{...}{nom des couleurs OFB (\link{ofb_colors}) à récupérer}
}
\value{

}
\description{
Récupérer les codes HEX des couleurs OFB
}
