% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{ofb_pal}
\alias{ofb_pal}
\title{Créer une fonction pour interpoler les couleurs des palettes OFB}
\usage{
ofb_pal(palette = "principale", reverse = FALSE, ...)
}
\arguments{
\item{palette}{nom de l'une des 8 palettes OFB (\link{ofb_palettes})}

\item{reverse}{valeur logique indiquant si les couleurs de la palette doivent
être inversées}

\item{...}{arguments supplémentaires à passer à \code{\link[=colorRampPalette]{colorRampPalette()}}}
}
\value{

}
\description{
Créer une fonction pour interpoler les couleurs des palettes OFB
}
