% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ofb_palettes}
\alias{ofb_palettes}
\title{Palettes de couleurs OFB}
\format{
liste composée de 8 palettes:
\itemize{
\item eau: dégradé de bleus
\item vegetal: dégradé de verts
\item terre: dégradé de marrons
\item terrestre: combinaison de verts et de marrons
\item feu: combinaison de rouge, orange et jaune
\item secheresse: combinaison de bleu, vert, orange et rouge
\item principale: combinaison des couleurs fortes (bleu, vert, jaune, orange et rouge)
\item pastel: combinaison de couleurs moins intenses (bleu, vert, jaune, orange et rouge)
}
}
\usage{
ofb_palettes
}
\description{
Les palettes sont constituées des couleurs (\link{ofb_colors}) de la charte graphique de l'OFB, regroupées par thématiques.
}
\keyword{datasets}
