% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{scale_color_ofb}
\alias{scale_color_ofb}
\title{Créer une fonction 'scale' pour les couleurs OFB}
\usage{
scale_color_ofb(palette = "principale", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{nom de l'une des 8 palettes OFB (\link{ofb_palettes})}

\item{discrete}{valeur logique indiquant si les couleurs sont discrètes ou non}

\item{reverse}{valeur logique indiquant si les couleurs de la palette doivent
être inversées}

\item{...}{arguments supplémentaires à passer à \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} ou \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}
}
\value{

}
\description{
Créer une fonction 'scale' pour les couleurs OFB
}
\examples{
\dontrun{
data <- data.frame(
        x = rnorm(100),
        y = rnorm(100),
        c = sample(c("a", "b", "c"), size = 100, replace = TRUE),
        d = rnorm(100)
      )

gg <- data \%>\%
    ggplot2::ggplot() +
    ggplot2::geom_point(ggplot2::aes(x = x, y = y, colour = d)) +
    ggplot2::labs(title = "titre", subtitle = "sous-titre")

gg
gg +
  scale_color_ofb(palette = "feu", discrete = FALSE)

gg2 <- data \%>\%
    ggplot2::ggplot() +
    ggplot2::geom_point(ggplot2::aes(x = x, y = y, colour = c)) +
    ggplot2::labs(title = "titre", subtitle = "sous-titre")

gg2
gg2 +
  scale_color_ofb(palette = "terrestre")
}
}
