% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{scale_fill_ofb}
\alias{scale_fill_ofb}
\title{Créer une fonction 'scale' pour les couleurs de remplissage  OFB}
\usage{
scale_fill_ofb(palette = "principale", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{nom de l'une des 8 palettes OFB (\link{ofb_palettes})}

\item{discrete}{valeur logique indiquant si les couleurs sont discrètes ou non}

\item{reverse}{valeur logique indiquant si les couleurs de la palette doivent
être inversées}

\item{...}{arguments supplémentaires à passer à \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} ou \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}}
}
\value{

}
\description{
Créer une fonction 'scale' pour les couleurs de remplissage  OFB
}
