% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_ofb}
\alias{theme_ofb}
\title{Thème ggplot OFB}
\usage{
theme_ofb(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{taille de la police utilisée (en points)}

\item{base_family}{type de police à utiliser}
}
\description{
Ce thème permet d'appliquer un visuel conforme à la charte graphique de l'OFB
à un graphique obtenu avec le package \{ggplot2\}
}
\examples{
\dontrun{
gg <- data.frame(
        x = rnorm(100),
        y = rnorm(100),
        c = sample(c("a", "b", "c"), size = 100, replace = TRUE),
        d = rnorm(100)
      ) \%>\%
    ggplot2::ggplot() +
    ggplot2::geom_point(ggplot2::aes(x = x, y = y, colour = c, size = d)) +
    ggplot2::labs(title = "titre", subtitle = "sous-titre")

gg
gg + theme_ofb()
}

}
